digital-watermarking-api/api/requirements.txt
Pillow
Flask
celery
opencv-python-headless
numpy
redis
pytesseract
scipy
flower
invisible-watermark